Finished css, adjusted delete calls
- Site is now fully responsive
- User age is now calculated
- Added state to user info
- Fixed footer issues
- Product deletes now checks if user is the same user of the product or is admin
- When deleting an user their product are also deleted

+++ NodeJS/Controllers/productController.js
@@ -84,14 +84,38 @@ router.delete('/:userId/:productId', verifyToken, (req,res)=>{
     if(!isValidObjectId(req.params.productId))
         console.log('No record with given id');
     else{
-        Product.findByIdAndRemove(req.params.productId, (err,docs)=>{
-            if(!err)
-              res.send(docs);
+        Product.findById(req.params.productId, (err,docs)=>{
+            if(!err){
+                if(docs.sellerId == req.params.userId){
+                    Product.findByIdAndRemove(req.params.productId, (err,docs)=>{
+                        if(!err)
+                          res.send(docs);
+                        else
+                          console.log('Error in deleting product with given id: ' + req.params.productId);
+                    }); 
+                }
+                else{
+                    User.findById(req.params.userId, (err,userResponse)=>{
+                        if(!err){
+                            if(userResponse.role == "ADMIN"){
+                                Product.findByIdAndRemove(req.params.productId, (err,docs)=>{
+                                    if(!err)
+                                      res.send(docs);
+                                    else
+                                      console.log('Error in deleting product with given id: ' + req.params.productId);
+                                }); 
+                            }
+                        }
+                        else{
+                            console.log("Error use is not admin or seller");
+                        }
+                    })
+                }
+            }
             else
-              console.log('Error in deleting product with given id: ' + req.params.productId);
-        });
+                console.log('Error in retrieving product with the given id: ' + req.params.id);
+        })
     }
-
 });
 
 module.exports = router;
